refactor: rename types.ts to index.ts(changing imports)
