+++ src/components/pages/main/products/controller/controller.ts
@@ -1,5 +1,5 @@
 import AppLoader from './appLoader';
-import { ProductType } from '../../../../types/types';
+import { ProductType } from '../../../../types';
 
 class AppController extends AppLoader {
   getSources(callback: ((data?: Array<ProductType>) => void) | undefined) {
